requirements.txt
lyft_dataset_sdk
flake8
numpy>=1.14.5
opencv-python>=3.4.2.17
pandas
Pillow>=5.2.0
pyquaternion>=0.9.5
scikit-image
scikit-learn>=0.19.2
tqdm>=4.25.0
scipy>=1.1.0
cachetools>=3.1.0
Shapely>=1.6.4.post2
fire
pytest
black
matplotlib
albumentations
ipdb
torchvision
torchsummary
tensorflow=1.14.0 
tensorboard
segmentation_models_pytorch
pytorch_toolbelt





